# Tude-Web-App-master/backend/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Étude+ API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /professors:
    get:
      operationId: listProfessors
      tags: [professors]
      summary: List professors
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: subject
          in: query
          schema:
            type: string
        - name: city
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: number
        - name: limit
          in: query
          schema:
            type: number
      responses:
        "200":
          description: Professors list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProfessorsListResponse"
  /professors/{id}:
    get:
      operationId: getProfessor
      tags: [professors]
      summary: Get professor profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Professor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Professor"
  /professors/{id}/approve:
    post:
      operationId: approveProfessor
      tags: [professors]
      summary: Approve professor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Professor approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Professor"
  /professors/{id}/reject:
    post:
      operationId: rejectProfessor
      tags: [professors]
      summary: Reject professor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Professor rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Professor"
...
  /stats/professor/{id}:
    get:
      operationId: getProfessorStats
      tags: [professors]
      summary: Get professor stats
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Professor stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProfessorStats"
components:
  schemas:
...
    Professor:
      type: object
      properties:
        id:
          type: number
        userId:
          type: number
        fullName:
          type: string
        profilePhoto:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        subjects:
          type: array
          items:
            type: string
        gradeLevels:
          type: array
          items:
            type: string
        yearsOfExperience:
          type: number
          nullable: true
        bio:
          type: string
          nullable: true
        qualifications:
          type: string
          nullable: true
        status:
          type: string
          enum: [pending, approved, rejected]
        rating:
          type: number
          nullable: true
        totalReviews:
          type: number
        totalStudents:
          type: number
        isVerified:
          type: boolean
        createdAt:
          type: string
      required: [id, userId, fullName, subjects, gradeLevels, status, totalReviews, totalStudents, isVerified, createdAt]
...
    ProfessorStats:
      type: object
      properties:
        totalEarnings:
          type: number
        platformFees:
          type: number
        totalStudents:
          type: number
        totalClasses:
          type: number
        pendingGrading:
          type: number
        averageRating:
          type: number
          nullable: true
        earningsByMonth:
          type: array
          items:
            $ref: "#/components/schemas/MonthlyEarning"
      required: [totalEarnings, platformFees, totalStudents, totalClasses, pendingGrading, earningsByMonth]
    MonthlyEarning:
      type: object
      properties:
        month:
          type: string
        amount:
          type: number
      required: [month, amount]
...
    ProfessorsListResponse:
      type: object
      properties:
        professors:
          type: array
          items:
            $ref: "#/components/schemas/Professor"
        total:
          type: number
        page:
          type: number
        limit:
          type: number
      required: [professors, total, page, limit]
...